fix: 인터페이스에 맞게 서비스 로직 호출 방식 수정
인터페이스 수정 사항에 맞게 Answers 라우트 핸들러에서 서비스 로직 호출 방식을 수정하였음

feat-#18

+++ src/apis/answers/answers.controller.ts
@@ -24,7 +24,7 @@ export class AnswersController {
 		@Req() req: Request & IAuthUser, //
 		@Body() createAnswerDTO: CreateAnswerDTO, //
 	): Promise<Answer> {
-		return this.answersService.createAnswer({ createAnswerDTO });
+		return this.answersService.createAnswer({ userId: req.user.id, createAnswerDTO });
 	}
 
 	/**
@@ -41,7 +41,7 @@ export class AnswersController {
 		@Param('id', ParseIntPipe) id: number, //
 		@Body() updateAnswerDTO: UpdateAnswerDTO,
 	): Promise<Answer> {
-		return this.answersService.updateAnswer({ id, updateAnswerDTO });
+		return this.answersService.updateAnswer({ userId: req.user.id, id, updateAnswerDTO });
 	}
 
 	/**
@@ -55,6 +55,6 @@ export class AnswersController {
 		@Req() req: Request & IAuthUser, //
 		@Param('id', ParseIntPipe) id: number, //
 	): Promise<void> {
-		return this.answersService.deleteAnswer({ id });
+		return this.answersService.deleteAnswer({ userId: req.user.id, id });
 	}
 }
